refactor: remove a lot of unused stuff
