+++ frontend/src/scenes/AuthenticatedShell.tsx
@@ -8,8 +8,6 @@ import { ToastCloseButton } from 'lib/lemon-ui/LemonToast/LemonToast'
 import { apiStatusLogic } from 'lib/logic/apiStatusLogic'
 import { eventIngestionRestrictionLogic } from 'lib/logic/eventIngestionRestrictionLogic'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
-import { PostOnboardingModal } from 'scenes/onboarding/PostOnboardingModal'
-import { postOnboardingModalLogic } from 'scenes/onboarding/postOnboardingModalLogic'
 import { WizardProgressFab } from 'scenes/onboarding/sdks/OnboardingInstallStep/WizardProgressFab'
 
 import { GlobalModals } from '~/layout/GlobalModals'
@@ -27,7 +25,6 @@ export default function AuthenticatedShell({ children }: { children: React.React
     useMountedLogic(eventIngestionRestrictionLogic)
     useMountedLogic(breadcrumbsLogic)
     useMountedLogic(globalSetupLogic)
-    useMountedLogic(postOnboardingModalLogic)
     useSetupHighlight()
 
     const { sceneConfig } = useValues(sceneLogic)
@@ -41,7 +38,6 @@ export default function AuthenticatedShell({ children }: { children: React.React
                 <GlobalModals />
                 <GlobalShortcuts />
                 <Command />
-                <PostOnboardingModal />
                 <ImpersonationNotice />
                 <SelfReadOnlyNotice />
                 <WizardProgressFab />
