+++ frontend/src/scenes/onboarding/productSelection/ProductSelectionShell.tsx
@@ -1,0 +1,13 @@
+import { OnboardingExitModal } from '../exit'
+import { ProductSelection } from './ProductSelection'
+
+export function ProductSelectionShell(): JSX.Element {
+    return (
+        <>
+            <ProductSelection />
+            <OnboardingExitModal />
+        </>
+    )
+}
+
+export default ProductSelectionShell
