--- frontend/src/scenes/onboarding/productSelection/variants/multiproduct/MultiproductProductSelection.tsx
@@ -1,5 +1,0 @@
-import { ProductCarousel } from '../ProductCarousel'
-
-export function MultiproductProductSelection(): JSX.Element {
-    return <ProductCarousel mode="multi" recommendationSource="multiproduct" />
-}
