--- frontend/src/scenes/onboarding/productSelection/variants/spotlight/SpotlightProductSelection.scss
@@ -1,48 +1,0 @@
-.SimplifiedProductSelection {
-    &__buzz {
-        animation: SimplifiedProductSelection__buzz 0.12s infinite linear;
-    }
-
-    @keyframes SimplifiedProductSelection__buzz {
-        0%,
-        100% {
-            transform: translate(0, 0) rotate(0deg);
-        }
-
-        10% {
-            transform: translate(-3px, 1px) rotate(-1.5deg);
-        }
-
-        20% {
-            transform: translate(3px, -1px) rotate(1.5deg);
-        }
-
-        30% {
-            transform: translate(-2px, 3px) rotate(-1deg);
-        }
-
-        40% {
-            transform: translate(2px, -3px) rotate(1deg);
-        }
-
-        50% {
-            transform: translate(-3px, 0) rotate(-1.5deg);
-        }
-
-        60% {
-            transform: translate(3px, 2px) rotate(1.5deg);
-        }
-
-        70% {
-            transform: translate(-2px, -2px) rotate(-1deg);
-        }
-
-        80% {
-            transform: translate(2px, 3px) rotate(1deg);
-        }
-
-        90% {
-            transform: translate(-3px, -3px) rotate(-1.5deg);
-        }
-    }
-}
