--- frontend/src/scenes/onboarding/productSelection/variants/spotlight/SpotlightProductSelection.tsx
@@ -1,5 +1,0 @@
-import { ProductCarousel } from '../ProductCarousel'
-
-export function SpotlightProductSelection(): JSX.Element {
-    return <ProductCarousel mode="single" recommendationSource="simplified" />
-}
